Remove glm dependencies in installed lib
